Updated plugin with release early configuration
