+++ Jenkinsfile
@@ -1,4 +1,28 @@
-node("slave-sbt") {
-  echo 'Hello World'
-}
+pipeline {
+    agent none
 
+    stages {
+        stage("Review") {
+            when {
+                expression { env.CHANGE_ID != null }
+            }
+            steps {
+                node("slave-sbt") {
+                    checkout scm
+                    sh 'sbt clean scalafmtSbtCheck compile'
+                }
+            }
+        }
+        stage("Release") {
+            when {
+                expression { env.CHANGE_ID == null }
+            }
+            steps {
+                node("slave-sbt") {
+                    checkout scm
+                    sh 'sbt clean releaseEarly'
+                }
+            }
+        }
+    }
+}
