+++ README.md
@@ -5,7 +5,7 @@
 ## SBT Nexus Plugin
 
 The plugin provides a collection of sbt auto plugins that configure a project with reasonable defaults for most sbt
-based scala projects.
+based scala nexus projects.
 
 Please visit the [parent project](https://github.com/BlueBrain/nexus) for more information about Nexus.
 
@@ -23,22 +23,6 @@ Add the following line to your `project/plugins.sbt` file:
 addSbtPlugin("ch.epfl.bluebrain.nexus" % "sbt-nexus" % "M.m.p")
 ```
 
-If you'd like to use this plugin to publish generated docker images add the following environment variables:
-```
-export DOCKER_REGISTRY="my-registry.com/my-project" # optional
-export DOCKER_BUILD_ARGS="HTTP_PROXY='http://my-proxy:80'|HTTPS_PROXY='http://my-proxy:80'|no_proxy='localhost,127.0.0.1'" #optional
-```
-
-If you're using macOS adding the above listed environment variables in your `~/.bash_profile` will not work for graphic
-applications (like IntelliJ, for instance). If you're using an ide, you should also set these values using `launchctl`
-(syntax: `launchctl setenv MYPATH myvar`), like so:
-```
-launchctl setenv DOCKER_REGISTRY "my-registry.com/my-project"
-launchctl setenv DOCKER_BUILD_ARGS "HTTP_PROXY='http://my-proxy:80'|HTTPS_PROXY='http://my-proxy:80'|no_proxy='localhost,127.0.0.1'"
-```
-
-**Note:** Environment variables that support multiple values are strings separated by '|'.
-
 ### Plugins
 
 #### CompilationPlugin
@@ -54,58 +38,16 @@ Exposed setting keys:
    * `scalacStrictFlags` (_scalac-strict-flags_): a collection of stricter compilation flags (i.e.: `-Xfatal-warnings`)
    * `scalacOptionalFlags` (_scalac-optional-flags_): useful additional flags (i.e.: `-Ypartial-unification`)
 
-The plugins automatically sets the `scalaVersion` to `2.12.3`, the `javaSpecificationVersion` to `1.8` and checks during
+The plugins automatically sets the `scalaVersion` to `2.12.4`, the `javaSpecificationVersion` to `1.8` and checks during
 the initialization phase if the installed jdk is compatible with the target java version.  It also appends to the
 `scalacOptions` all the flags defined in the exposed settings and sets the necessary compiler flags for both `javac` and
 `scalac` to generate bytecode compatible with the defined `javaSpecificationVersion`.
 
 #### ReleasePlugin
 
-The `ReleasePlugin` configures the release process with explicit steps and computes the release and next versions based
-on environment variables (if present).
+The `ReleasePlugin` configures the release process using for `sbt-release-early` plugin.
 
-Issue the following command to initiate a release: `sbt 'release with-defaults'`
-
-The configured release process steps are configured as follows:
-   1. check the working dir is clean
-   2. check the project has no snapshot dependencies
-   3. compute the appropriate release and next development version
-   4. clean the project
-   5. run tests (implies compile)
-   6. update the `version.sbt` with the release version
-   7. commit the changes to `version.sbt`
-   8. tag the release with __$artifactId-$version__
-   9. publish the artifact according to the publish settings
-   10. update the `version.sbt` with the next development version
-   11. commit the changes to `version.sbt`
-   12. push the changes to the remote git repository
-
-By default, the current configuration releases the artifact using the version defined in `version.sbt` with its
-__-SNAPSHOT__ qualifier stripped.  The next version is computed by incrementing the patch (bugfix) segment of the
-release version.  For instance, if the current version is __0.3.5-SNAPSHOT__ the release version will be set to
-__0.3.5__ and the next development version will be set to __0.3.6-SNAPSHOT__.
-
-In order to change the default computed versions, for instance in situations where you need to release a new minor or
-major version, you can pass to SBT the following environment variables:
-
-   * `RELEASE_VERSION`: the exact version to release (__note:__ if it contains a qualifier it will be stripped)
-   * `NEXT_VERSION`: the next development version to be added to the `version.sbt` file
-
-Both of these variables are optional.  For example:
-
-```shell
-$ sbt 'release with-defaults'
-$ RELEASE_VERSION="0.4.0" sbt 'release with-defaults'
-$ RELEASE_VERSION="0.4.0" NEXT_VERSION="0.5.0-SNAPSHOT" sbt 'release with-defaults'
-$ NEXT_VERSION="0.5.0-SNAPSHOT" sbt 'release with-defaults'
-```
-
-#### PublishPlugin
-
-The `PublishPlugin` configures the publishing step, currently allowing to blacklist dependencies in the resulting pom
-file.
-
-Exposed setting keys:
+Additional exposed setting keys:
 
    * `dependencyBlacklist` (_dependency-blacklist_): a module filter for stripping out compile time only dependencies
      from the resulting pom file.
@@ -113,6 +55,7 @@ Exposed setting keys:
 The plugin hooks into the `pomPostProcess` task and removes the _dependency_ xml nodes that match the filter defined by
 the `dependencyBlacklist`.  The filter is currently configured to strip out coverage and static analysis dependencies
 (_scoverage_ and _scapegoat_).
+
 
 #### StaticAnalysisPlugin
 
@@ -162,47 +105,9 @@ lazy val myProject = project.settings(Seq(
 
 The default configuration automatically adds the _scala library_ documentation url to the _apiMappings_.
 
-#### ServicePackagingPlugin
+#### InstrumentationPlugin
 
-The `ServicePackagingPlugin` automates the configuration for creating and publishing docker images.  The auto plugin
-needs to be enabled _explicitly_ along with its dependencies, i.e.:
-
-```scala
-lazy val myProject = project
-  .enablePlugins(UniversalPlugin, JavaAppPackaging, DockerPlugin, ServicePackagingPlugin)
-  .settings(packageName in Docker := "my-image-name")
-```
-
-Enabling the plugin also requires setting the appropriate image name, as in the example above.
-
-The docker image is being built referencing `openjdk:8-jre` as the parent image.
-
-The plugin also hooks into the `publishLocal` and `publish` tasks to automatically publish the produced docker image
-to the configured repository.
-
-Please see the [sbt native packager documentation](http://www.scala-sbt.org/sbt-native-packager/) for more information
-about possible configuration options.
-
-#### DocsPackagingPlugin
-
-The `DocsPackagingPlugin` automates the configuration for generating and publishing docker images containing
-documentation generated with [paradox](https://github.com/lightbend/paradox).
-
-The generated image uses an nginx to serve the generated static content and describes two environment variables
-(`LOCATION` and `SERVER_NAME`) that can be customized in the nginx config.
-
-The following nginx configuration template is used:
-```
-server {
-    listen 8080;
-    server_name ${SERVER_NAME};
-    port_in_redirect off;
-    location ${LOCATION} {
-        alias   /usr/share/nginx/html/;
-        index  index.html;
-    }
-}
-```
+This plugin adds aspectj and sigar loader as java agents to all apps packaged with universal packaging.
 
 #### NexusPlugin
