+++ project/plugins.sbt
@@ -1,3 +1,3 @@
-addSbtPlugin("com.github.gseitz" % "sbt-release"  % "1.0.7")
-addSbtPlugin("com.geirsson"      % "sbt-scalafmt" % "1.4.0")
-addSbtPlugin("org.foundweekends" % "sbt-bintray"  % "0.5.3")
+addSbtPlugin("com.geirsson"    % "sbt-scalafmt"      % "1.4.0")
+addSbtPlugin("io.get-coursier" % "sbt-coursier"      % "1.0.1")
+addSbtPlugin("ch.epfl.scala"   % "sbt-release-early" % "2.1.1")
