+++ src/main/scala/ch/epfl/bluebrain/nexus/sbt/nexus/NexusPlugin.scala
@@ -15,6 +15,7 @@ object NexusPlugin extends AutoPlugin {
   override lazy val buildSettings = Seq(
     resolvers ++= Seq(
       Resolver.bintrayRepo("bbp", "nexus-releases"),
+      Resolver.bintrayRepo("bbp", "nexus-snapshots"),
     )
   )
 
@@ -22,6 +23,7 @@ object NexusPlugin extends AutoPlugin {
     organization := "ch.epfl.bluebrain.nexus",
     resolvers ++= Seq(
       Resolver.bintrayRepo("bbp", "nexus-releases"),
+      Resolver.bintrayRepo("bbp", "nexus-snapshots"),
     )
   )
 }
