+++ src/main/scala/ch/epfl/bluebrain/nexus/sbt/nexus/StaticAnalysisPlugin.scala
@@ -14,7 +14,7 @@ object StaticAnalysisPlugin extends AutoPlugin {
   override lazy val trigger = allRequirements
 
   override lazy val buildSettings = Seq(
-    scapegoatVersion     := "1.3.3",
+    scapegoatVersion     := "1.3.4",
     scapegoatMaxWarnings := 0,
     scapegoatMaxErrors   := 0,
     scapegoatMaxInfos    := 0,
